Add QueryClient to the options for useQuery, useMutation, etc
This is used in preference to taking it from the context, and
can be useful in unusual situations when it's difficult to get
the query client into or out of the Svelte context.

useQueries doesn't take a single options object, so it gets an
extra optional parameter to pass the client.